Implementação de DTO  terminada

+++ src/main/java/com/luiz/primeiroexemplo/services/ProdutoService.java
@@ -2,12 +2,16 @@ package com.luiz.primeiroexemplo.services;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.luiz.primeiroexemplo.model.Produto;
+import com.luiz.primeiroexemplo.model.exception.ResourceNotFoundException;
 import com.luiz.primeiroexemplo.repository.ProdutoRepository;
+import com.luiz.primeiroexemplo.shared.ProdutoDTO;
 
 @Service
 public class ProdutoService {
@@ -19,8 +23,11 @@ public class ProdutoService {
      * Retorna uma lista de produtos
      * @return Lista de produtos
      */
-    public List<Produto> obterTodos(){
-        return produtoRepository.findAll();
+    public List<ProdutoDTO> obterTodos(){
+        List<Produto> produtos = produtoRepository.findAll();
+        return produtos.stream()
+        .map(produto -> new ModelMapper().map(produtos, ProdutoDTO.class))
+        .collect(Collectors.toList());
     }
 
     /**
@@ -28,8 +35,14 @@ public class ProdutoService {
      * @param id do produto que sera localizado 
      * @return retorna um produto caso tenha encontrado
      */
-    public Optional<Produto> obterPorId(Integer id){
-        return produtoRepository.findById(id);
+    public Optional<ProdutoDTO> obterPorId(Integer id){
+        Optional<Produto> produto = produtoRepository.findById(id);
+
+        if(produto.isEmpty()){
+            throw new ResourceNotFoundException("Produto com o id: " + id + "não encontrado");
+        }
+        ProdutoDTO dto = new ModelMapper().map(produto.get(), ProdutoDTO.class);
+        return Optional.of(dto);
     }
 
     /**
@@ -37,8 +50,14 @@ public class ProdutoService {
      * @param produto que será adicionando
      * @return Retorna o produto que foi adicionado na lista
      */
-    public Produto adicionar(Produto produto){
-        return produtoRepository.save(produto);
+    public ProdutoDTO adicionar(ProdutoDTO produtoDto){
+
+        produtoDto.setId(null);
+        ModelMapper mapper = new ModelMapper();
+        Produto produto = mapper.map(produtoDto, Produto.class);
+        produto = produtoRepository.save(produto);
+        produtoDto.setId(produto.getId());
+        return produtoDto;
     }
 
     /**
@@ -46,6 +65,12 @@ public class ProdutoService {
      * @param id do produto a ser deletado
      */
     public void deletar(Integer id){
+        Optional<Produto> produto = produtoRepository.findById(id);
+
+        if(produto.isEmpty()){
+            throw new ResourceNotFoundException("Nap foi possivel deletar o produto com esse id: "+id+". Produto não existente.");
+        }
+
         produtoRepository.deleteById(id);
     }
 
@@ -55,9 +80,16 @@ public class ProdutoService {
      * @param id do produto
      * @returnretorna um produto apos atualizar a lista 
      */
-    public Produto atualizar(Integer id, Produto produto){
-        produto.setId(id);
-        return produtoRepository.save(produto);
+    public ProdutoDTO atualizar(Integer id, ProdutoDTO produtoDTO){
+        produtoDTO.setId(id);
+
+        ModelMapper mapper = new ModelMapper();
+        
+        Produto produto  = mapper.map(produtoDTO, Produto.class);
+
+        produtoRepository.save(produto);
+
+        return produtoDTO;
     }
 
 }
